+++ session.opencode.sh
@@ -1,0 +1,1 @@
+ opencode -s ses_17a0cfaeeffejscQ1f0eNYnpBn